Add base config UI and widen move timeframe
Add a base UI for the config screen on the main menu which includes 1) a secondary 1280 x 720 window dedicated to game customization, 2) a base frame for the player's avatar, 3) a game settings section to be implemented later, and 4) a close button to return to the main menu. Adjust movement timeframe to allow 5 extra frames for the player to click with score adjustment applied. Note that movements must be done in one-measure-length movement, meaning one movement every 4 beats in order to accurately display results. May change this later depending on time.