Change menu screen, add level 1 music, update rhythm engine
The rhythm engine is now based on solely the media soundtrack rather than frames or a timer and does calculations for beats by itself since this is the most accurate way of representing beats in a song. To change the rhythm to match any song, alter the "bpm" variable and the "OST" variable to match a song's BPM and song file respectively.