Added Configuration Screen
Added the base config options: difficulty, weapon select, and character name. However, it is very rough looking and the experience is not super smooth. The operation could be pretty easily ironed out but that's not my first priority right now.
The UI elements are kind of in the wrong places, I just took rough guesses at the positions. And this will really have to be revamped if we are allowing for multiple resolutions.
Speaking of which, the options which Kyle presented on the concept art screen for the game configuration would probably be best for the pause menu